Use common function for authenticating admin requests
- move admin API error functions
